backend/requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
python-jose[cryptography]
